requirements.txt
selenium~=4.33.0
webdriver-manager~=4.0.2
pytest~=7.4.0
pyyaml~=6.0
pytest-html
